Move SendRequest and NetworkMessageSender to crate::peer
With the integration of the PeerManager and Interconnect
these structs are being moved to their final location.
The Network struct is being deprecated, along with the
NetworkMessageQueue.
